Change AddressService methods [#20484]
The logic of the GetChildObjectsAsync and GetPathFromRootToObjectAsync functions has been changed, now they work more efficiently, due to less access to the database, and a bug has also been fixed when not all elements were found when searching by words. The House model was changed, the FullName field appeared in it, which stores the full name of the house, so the getFullName function was removed, since it is no longer needed. The values of the maximum number of elements have been moved to appsettings.