Fix the mistakes made in the previous development steps [#20804]
Added to the methods of the service, where the address is needed, a check of the Guid belonging to the buildings. I also hung the [Range(0, 10)] attribute on the parameter for evaluation, now it is impossible to specify an estimate outside the range from 0 to 10. Added a check to the CheckToSetRatingWithoutCheckDishAsync function of DishService to check that the order has been delivered. In the GetChildObjectsAsync method, duplicates of address elements are now excluded from the query result.