Refactoring code style and change ValidationProblemException
Removed TokenLogoutDTO, as it is no longer needed. A default value has been added to each request value from appsettings. Changed ValidationProblemException, now it accepts either key and message, or HttpValidationProblemDetails. Private class fields have been renamed, now they start with the _ sign. Also fixed an error with the evaluation of the dish, because of which it was impossible to put it.